requirements.txt
oauthlib==1.0.1
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
